+++ robotframework-ls/docs/changelog.md
@@ -24,6 +24,7 @@ NEXT
 
 - [VSCode] A non-string value is converted to string before expanding variables. [#727](https://github.com/robocorp/robotframework-lsp/issues/727)
 - [Intellij] Fixed NullPointerException on hover. [#731](https://github.com/robocorp/robotframework-lsp/issues/731)
+- [Intellij] `$Prompt$` macro properly replaced when launching. [#737](https://github.com/robocorp/robotframework-lsp/issues/737)
 - Operations no longer timeout, rather, they just print to the log (as the timeouts weren't always ideal for slower machines). [#733](https://github.com/robocorp/robotframework-lsp/issues/733)
 - Fixed issue where references wouldn't be found properly.